functions/requirements.txt
# Firebase and Cloud Functions libraries
firebase-functions
firebase-admin

# Google Cloud and Vertex AI libraries
google-cloud-aiplatform
google-cloud-firestore